# candidate-match-export/candidate-match/openapi.yaml
openapi: 3.0.3
info:
  title: Cloudberries Candidate Match API
  version: 1.1.0
  description: |
    API for candidate matching, CV retrieval, embeddings, and health checks.
    Includes normalized health responses and skill-based matching endpoint.
    Generated from controller sources as of this commit.
  contact:
    name: Cloudberries
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /skills:
    get:
      tags: [ Skills ]
      summary: List skills across consultants
      parameters:
        - in: query
          name: skill
          description: Optional skill filter (repeat to filter by multiple)
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
      responses:
        '200':
          description: Skill aggregates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SkillInCompanyDto'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /skills/names:
    get:
      tags: [ Skills ]
      summary: List all skill names
      parameters:
        - in: query
          name: prefix
          required: false
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
      responses:
        '200':
          description: Array of skill names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        default:
          $ref: '#/components/responses/ErrorResponse'
  /skills/summary:
    get:
      tags: [ Skills ]
      summary: Paged summary of skills with consultant counts
      parameters:
        - in: query
          name: q
          required: false
          schema: { type: string }
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - in: query
          name: sort
          required: false
          schema: { type: string, default: 'consultantCount,desc' }
      responses:
        '200':
          description: Paged skill summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      type: object
                      properties:
                        name: { type: string }
                        consultantCount: { type: integer }
                  number: { type: integer }
                  size: { type: integer }
                  totalElements: { type: integer }
                  totalPages: { type: integer }
                  first: { type: boolean }
                  last: { type: boolean }
                  sort: { type: object, additionalProperties: true }
                  pageable: { type: object, additionalProperties: true }
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /project-requests:
    get:
      tags: [ Project Requests ]
      summary: List stored customer project requests with pagination
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
          description: Page number (0-based)
        - in: query
          name: size
          schema:
            type: integer
            default: 20
          description: Number of items per page
        - in: query
          name: sort
          schema:
            type: string
            default: "id,desc"
          description: Sort specification (field,direction)
      responses:
        '200':
          description: Paged project requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedProjectRequestResponseDto'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [ Project Requests ]
      summary: Create a new project request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequestDto'
            example:
              customerName: "Acme Corp AS"
              requiredSkills: ["KOTLIN", "JAVA"]
              startDate: "2024-04-01T09:00:00"
              endDate: "2024-10-01T17:00:00"
              responseDeadline: "2024-03-15T17:00:00"
              requestDescription: "We need a senior backend developer for a 6-month project using Kotlin and Spring Boot."
              responsibleSalespersonEmail: "sales@acme.com"
      responses:
        '201':
          description: Project request created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRequestDto'
        '400':
          description: Invalid request data
        default:
          $ref: '#/components/responses/ErrorResponse'
  /project-requests/upload:
    post:
      tags: [ Project Requests ]
      summary: Upload and analyze a customer project request PDF
      description: |
        Uploads a PDF document and uses AI to extract structured information including:
        - Customer name from the document
        - Project summary and requirements
        - MUST vs SHOULD requirements categorization
        - Project deadline (if mentioned)
        - Upload timestamp tracking
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [ file ]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Extracted project request with AI-analyzed structure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRequestResponseDto'
              example:
                id: 123
                customerName: "Acme Corporation AS"
                title: "Senior Kotlin Developer - Mobile Platform"
                summary: "Project for developing a new mobile banking application using Kotlin and modern Android architecture. Requires experienced developer with strong backend integration skills."
                originalFilename: "project-request-acme-mobile.pdf"
                uploadedAt: "2024-10-01T10:30:00Z"
                deadlineDate: "2024-12-15T00:00:00Z"
                mustRequirements:
                  - name: "5+ years experience with Kotlin/Android development"
                    details: null
                  - name: "Experience with REST API integration"
                    details: null
                  - name: "Knowledge of Material Design principles"
                    details: null
                shouldRequirements:
                  - name: "Experience with Jetpack Compose"
                    details: null
                  - name: "Familiarity with CI/CD pipelines"
                    details: null
                  - name: "Previous fintech application experience"
                    details: null
        default:
          $ref: '#/components/responses/ErrorResponse'
  /project-requests/{id}:
    get:
      tags: [ Project Requests ]
      summary: Get stored customer project request by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Project request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRequestResponseDto'
        '404':
          description: Not found
        default:
          $ref: '#/components/responses/ErrorResponse'
  /project-requests/{id}/close:
    put:
      tags: [ Project Requests ]
      summary: Close a project request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Project request closed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRequestDto'
        '404':
          description: Project request not found
        default:
          $ref: '#/components/responses/ErrorResponse'
  /project-requests/{id}/analyze:
    post:
      tags: [ Project Requests ]
      summary: Trigger AI analysis for a project request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Analysis completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRequestDto'
        '404':
          description: Project request not found
        default:
          $ref: '#/components/responses/ErrorResponse'
  /project-requests/{id}/suggestions:
    get:
      tags: [ Project Requests ]
      summary: Get AI suggestions for a project request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: AI suggestions for the project request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AISuggestionDto'
        '404':
          description: Project request not found
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  parameters:
    PageParam:
      in: query
      name: page
      description: Page number (0-indexed)
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      in: query
      name: size
      description: Page size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    SkillInCompanyDto:
      type: object
      required: [ name, konsulenterMedSkill, konsulenter ]
      properties:
        name:
          type: string
        consultantCount:
          type: integer
          description: Number of consultants with this skill (preferred)
        konsulenterMedSkill:
          type: integer
          description: Number of consultants with this skill (deprecated)
          deprecated: true
        konsulenter:
          type: array
          items:
            $ref: '#/components/schemas/ConsultantSummaryDto'
...
    ConsultantSummaryDto:
      type: object
      required: [ userId, name, email, bornYear, defaultCvId ]
      properties:
        userId:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        bornYear:
          type: integer
        defaultCvId:
          type: string
...
    ProjectRequestResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customerName:
          type: string
          description: Customer name extracted from the document
        title:
          type: string
          description: Title/subject derived from the document
        summary:
          type: string
          description: AI-generated summary of the project request
        originalFilename:
          type: string
          description: Original filename of the uploaded PDF
        uploadedAt:
          type: string
          format: date-time
          description: Timestamp when the document was uploaded
        deadlineDate:
          type: string
          format: date-time
          description: Project deadline extracted from the document (if available)
        mustRequirements:
          type: array
          items:
            $ref: '#/components/schemas/ProjectRequirementDto'
          description: Critical requirements that must be met
        shouldRequirements:
          type: array
          items:
            $ref: '#/components/schemas/ProjectRequirementDto'
          description: Preferred requirements that would be beneficial
    ProjectRequirementDto:
      type: object
      required: [ name ]
      properties:
        name:
          type: string
        details:
          type: string
    PagedProjectRequestResponseDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProjectRequestResponseDto'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        currentPage:
          type: integer
        pageSize:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
    CreateProjectRequestDto:
      type: object
      required: [ customerName, requiredSkills, startDate, endDate, responseDeadline, requestDescription, responsibleSalespersonEmail ]
      properties:
        customerName:
          type: string
          minLength: 1
        requiredSkills:
          type: array
          items:
            type: string
            enum: [ KOTLIN, JAVA, PYTHON, JAVASCRIPT, TYPESCRIPT, REACT, ANGULAR, VUE, SPRING_BOOT, BACKEND, FRONTEND, FULLSTACK, AZURE, AWS, DOCKER, KUBERNETES ]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        responseDeadline:
          type: string
          format: date-time
        status:
          type: string
          enum: [ OPEN, IN_PROGRESS, CLOSED ]
          default: OPEN
        requestDescription:
          type: string
          minLength: 1
        responsibleSalespersonEmail:
          type: string
          format: email
    ProjectRequestDto:
      type: object
      required: [ customerName, requiredSkills, startDate, endDate, responseDeadline, status, requestDescription, responsibleSalespersonEmail ]
      properties:
        id:
          type: integer
          format: int64
        customerId:
          type: integer
          format: int64
        customerName:
          type: string
        requiredSkills:
          type: array
          items:
            type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        responseDeadline:
          type: string
          format: date-time
        status:
          type: string
          enum: [ OPEN, IN_PROGRESS, CLOSED ]
        requestDescription:
          type: string
        responsibleSalespersonEmail:
          type: string
          format: email
        aiSuggestions:
          type: array
          items:
            $ref: '#/components/schemas/AISuggestionDto'
    AISuggestionDto:
      type: object
      required: [ consultantName, userId, cvId, matchScore, justification, createdAt ]
      properties:
        id:
          type: integer
          format: int64
        consultantName:
          type: string
        userId:
          type: string
        cvId:
          type: string
        matchScore:
          type: number
          format: double
          minimum: 0.0
          maximum: 100.0
        justification:
          type: string
        createdAt:
          type: string
          format: date-time
        skills:
          type: array
          items:
            type: string
...
    Problem:
      type: object
      description: Generic error payload
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

    # AI Chat Search Schemas
...